backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-socketio
aiofiles
pydantic
pydantic-settings
opencv-python
numpy